+++ typescript-client-sdk/src/sdk/sdk.ts
@@ -48,7 +48,7 @@ export class SDK {
   public _serverURL: string;
   private _language = "typescript";
   private _sdkVersion = "0.1.2";
-  private _genVersion = "2.26.2";
+  private _genVersion = "2.26.3";
   private _globals: any;
 
   constructor(props?: SDKProps) {